refactored GetAllBeaonsService to remove mapping logic
