+++ src/main/java/uk/gov/mca/beacons/service/dto/BeaconDTO.java
@@ -1,0 +1,17 @@
+package uk.gov.mca.beacons.service.dto;
+
+import java.util.UUID;
+
+public class BeaconDTO {
+
+    private UUID id;
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+}
