+++ src/test/java/uk/gov/mca/beacons/service/registrations/BeaconsControllerIntegrationTest.java
@@ -118,6 +118,28 @@ class BeaconsControllerIntegrationTest {
       .exists();
   }
 
+  @Test
+  void requestBeaconControllerShouldReturnBeaconByUuidOther() {
+    String uuidAsString = uuid.toString();
+    var request = makeGetRequest(String.format("/beacons/other/%s", uuidAsString));
+
+     request.jsonPath("$.id").exists();
+
+    // request.jsonPath("$.meta.pageSize").exists();
+    // request.jsonPath("$.meta.count").exists();
+    // request.jsonPath("$.data").exists();
+    // request.jsonPath("$.data[0].type").isEqualTo("beacon");
+    // request.jsonPath("$.data[0].id").isEqualTo(uuidAsString);
+    // request.jsonPath("$.data[0].attributes.hexId").exists();
+    // request.jsonPath("$.data[0].attributes.manufacturer").exists();
+    // request.jsonPath("$.data[0].attributes.status").exists();
+    // request.jsonPath("$.data[0].attributes.uses[0].environment").exists();
+    // request.jsonPath("$.data[0].attributes.owner.fullName").exists();
+    // request
+    //   .jsonPath("$.data[0].attributes.emergencyContacts[0].fullName")
+    //   .exists();
+  }
+
   private WebTestClient.BodyContentSpec makeGetRequest(String url) {
     return webTestClient
       .get()
